LayoutTests/accessibility/mac/press-action-for-presentational-descendants.html

<!DOCTYPE HTML>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>

<div id="content" role="group">

<div role="presentation" onclick="handleClick(this);">
  <button>this is a button (user assumes clickable)</button>
  <span>this is no longer exposed as clickable because of the explicit role on the delegate</span>
</div>

</div>

<div id="console"></div>
<script>

description("This tests ensures that static text that is inside a click handler parent who IS presentational will not say it supports press.")

if (window.testRunner && window.accessibilityController) {

   var button = accessibilityController.accessibleElementById("content").childAtIndex(0);
   var text = accessibilityController.accessibleElementById("content").childAtIndex(1);
   shouldBeTrue("button.isPressActionSupported()");
   shouldBeFalse("text.isPressActionSupported()");
}

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
